+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Improved typescript typings
 - Removed ability to add custom request store due to limitations
 - Added ability to specify request codes and methods that should be retried
+- Added more lifecycle hooks
 
 ## 1.0.0-beta.1 - 24-12-2024